+++ src/libft.h
@@ -6,7 +6,7 @@
 /*   By: alrobert <alrobert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/26 17:26:54 by alrobert          #+#    #+#             */
-/*   Updated: 2022/11/10 17:46:10 by alrobert         ###   ########.fr       */
+/*   Updated: 2022/11/11 17:49:42 by alrobert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,9 +47,9 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strnstr(const char *big, const char *little, size_t len);
 void	ft_bzero(void *s, size_t n);
-void	ft_putnbr_fd(int n, int fd);
+void	ft_putnbr_fd(long long n, int fd);
 void	ft_putendl_fd(char *s, int fd);
-void	ft_puthexa(unsigned long int nbr, int is_maj_or_min);
+void	ft_puthexa(unsigned long nbr, int is_maj_or_min);
 void	*ft_memset(void *s, int c, size_t n);
 void	*ft_calloc(size_t nmemb, size_t size);
 void	ft_lstadd_back(t_list **lst, t_list *new);
@@ -64,9 +64,9 @@ void	ft_striteri(char *s, void (*f)(unsigned int, char*));
 t_list	*ft_lstlast(t_list *lst);
 t_list	*ft_lstnew(void *content);
 t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
-size_t	ft_nblen(int nb);
+size_t	ft_nblen(long nb);
 size_t	ft_strlen(const char *c);
-size_t	ft_memlen(unsigned long int nbr);
+size_t	ft_memlen(unsigned long nbr);
 size_t	ft_strlcat(char *dest, const char *src, size_t n);
 size_t	ft_strlcpy(char *dst, const char *src, size_t size);
